Rename rule to no-nested-assignment
