+++ eslint-bridge/src/rules/main.ts
@@ -41,7 +41,6 @@ import { rule as noRedundantJump } from "./no-redundant-jump";
 import { rule as cyclomaticComplexity } from "./cyclomatic-complexity";
 import { rule as noAlphabeticalSort } from "./no-alphabetical-sort";
 import { rule as noCollectionSizeMischeck } from "./no-collection-size-mischeck";
-import { rule as noConditionalAssignment } from "./no-conditional-assignment";
 import { rule as noTab } from "./no-tab";
 import { rule as noEmptyCollection } from "./no-empty-collection";
 import { rule as noTryPromise } from "./no-try-promise";
@@ -49,6 +48,7 @@ import { rule as nonExistentOperator } from "./non-existent-operator";
 import { rule as noInMisuse } from "./no-in-misuse";
 import { rule as noInvariantReturns } from "./no-invariant-returns";
 import { rule as noMsleadingArrayReverse } from "./no-misleading-array-reverse";
+import { rule as noNestedAssignment } from "./no-nested-assignment";
 import { rule as noNestedSwitch } from "./no-nested-switch";
 import { rule as noNestedIncDec } from "./no-nested-incdec";
 import { rule as noNestedTemplateLiterals } from "./no-nested-template-literals";
@@ -101,7 +101,6 @@ ruleModules["no-redundant-jump"] = noRedundantJump;
 ruleModules["cyclomatic-complexity"] = cyclomaticComplexity;
 ruleModules["no-alphabetical-sort"] = noAlphabeticalSort;
 ruleModules["no-collection-size-mischeck"] = noCollectionSizeMischeck;
-ruleModules["no-conditional-assignment"] = noConditionalAssignment;
 ruleModules["no-tab"] = noTab;
 ruleModules["no-empty-collection"] = noEmptyCollection;
 ruleModules["no-try-promise"] = noTryPromise;
@@ -109,6 +108,7 @@ ruleModules["non-existent-operator"] = nonExistentOperator;
 ruleModules["no-in-misuse"] = noInMisuse;
 ruleModules["no-invariant-returns"] = noInvariantReturns;
 ruleModules["no-misleading-array-reverse"] = noMsleadingArrayReverse;
+ruleModules["no-nested-assignment"] = noNestedAssignment;
 ruleModules["no-nested-switch"] = noNestedSwitch;
 ruleModules["no-nested-incdec"] = noNestedIncDec;
 ruleModules["no-nested-template-literals"] = noNestedTemplateLiterals;
