+++ eslint-bridge/src/rules/no-nested-assignment.ts
