+++ javascript-checks/src/main/java/org/sonar/javascript/checks/CheckList.java
@@ -66,7 +66,6 @@ public final class CheckList {
       ArrayCallbackWithoutReturnCheck.class,
       ArrayMutatingOperationCheck.class,
       ArrowFunctionConventionCheck.class,
-      AssignmentWithinConditionCheck.class,
       AssociativeArraysCheck.class,
       BackboneChangedIsUsedCheck.class,
       BitwiseOperatorsCheck.class,
@@ -176,6 +175,7 @@ public final class CheckList {
       MultilineBlockCurlyBraceCheck.class,
       MultilineStringLiteralsCheck.class,
       NamedFunctionExpressionCheck.class,
+      NestedAssignmentCheck.class,
       NestedConditionalOperatorsCheck.class,
       NestedControlFlowDepthCheck.class,
       NewOperatorMisuseCheck.class,
